pyproject.toml
[project]
name = "osd-text-extractor"
version = "0.1.0"
description = "Module to extract plain text from different formats for LLM purposes"
authors = [
    {name = "OneSlap Team", email = "lxarbuz@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
keywords = ["text-extraction", "pdf", "docx", "xlsx", "html", "llm", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Office Suites",
    "Typing :: Typed",
]

dependencies = [
    "beautifulsoup4>=4.13.5",
    "dishka>=1.6.0",
    "emoji>=2.14.1",
    "lxml>=6.0.1",
    "odfpy>=1.4.1",
    "openpyxl>=3.1.5",
    "pymupdf>=1.26.4",
    "python-docx>=1.2.0",
    "striprtf>=0.0.29",
]

[project.optional-dependencies]
dev = [
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
    "build>=1.2.0",
    "twine>=6.0.0",
    # Линтеры и форматтеры
    "bandit>=1.8.6",
    "mypy>=1.17.1",
    "pylint>=3.3.8",
    "ruff>=0.6.9,<0.7.0",
    # Тестирование
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    # Безопасность
    "safety>=3.2.8",
    # Pre-commit
    "pre-commit>=3.8.0",
]

[project.urls]
"Homepage" = "https://github.com/OneSlap/osd-text-extractor"
"Bug Reports" = "https://github.com/OneSlap/osd-text-extractor/issues"
"Source" = "https://github.com/OneSlap/osd-text-extractor"
"Documentation" = "https://github.com/OneSlap/osd-text-extractor#readme"

[build-system]
requires = ["setuptools>=80.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "D",   # pydocstyle
    "S",   # flake8-bandit
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "ERA", # eradicate
]
ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "S101", # Use of assert detected
    "N999", # Invalid module name
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D", "ARG001"]
"src/osd_text_extractor/__init__.py" = ["F401"]
"src/osd_text_extractor/*/exceptions.py" = ["N818"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict = true
exclude = [
    "tests/",
    "build/",
    "htmlcov/",
]

[[tool.mypy.overrides]]
module = [
    "docx.*",
    "fitz.*",
    "odf.*",
    "openpyxl.*",
    "striprtf.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--disable-warnings",
    "--cov=src/osd_text_extractor",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html",
    "--cov-fail-under=80"
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "default",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/osd_text_extractor"]
omit = [
    "**/__init__.py",
    "**/tests/**",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\(Protocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "build", "htmlcov"]
skips = ["B101"]  # Skip assert_used test

[tool.pylint.main]
source-roots = ["src"]
py-version = "3.12"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "import-error",
]

[tool.pylint.format]
max-line-length = 88
